+++ liblzma-rs/src/common/block_buffer_encoder.rs
@@ -207,7 +207,7 @@ pub const LZMA_VLI_BYTES_MAX: c_int = 9 as c_int;
 pub const LZMA_CHECK_ID_MAX: c_int = 15 as c_int;
 pub const LZMA_CHECK_SIZE_MAX: c_int = 64 as c_int;
 pub const LZMA_FILTER_LZMA2: c_ulonglong = 0x21 as c_ulonglong;
-pub const LZMA_DICT_SIZE_MIN: c_uint = 4096 as c_uint;
+pub const LZMA_DICT_SIZE_MIN: c_uint = 4096;
 pub const LZMA_BLOCK_HEADER_SIZE_MAX: c_int = 1024 as c_int;
 pub const COMPRESSED_SIZE_MAX: c_ulonglong = LZMA_VLI_MAX
     .wrapping_sub(LZMA_BLOCK_HEADER_SIZE_MAX as c_ulonglong)
@@ -295,7 +295,7 @@ unsafe extern "C" fn block_encode_uncompressed(
     filters[1 as usize].id = LZMA_VLI_UNKNOWN as lzma_vli;
     let mut filters_orig: *mut lzma_filter = (*block).filters;
     (*block).filters = &raw mut filters as *mut lzma_filter;
-    if lzma_block_header_size(block) as c_uint != LZMA_OK as c_uint {
+    if lzma_block_header_size(block) != LZMA_OK {
         (*block).filters = filters_orig;
         return LZMA_PROG_ERROR;
     }
@@ -305,7 +305,7 @@ unsafe extern "C" fn block_encode_uncompressed(
         (*block).filters = filters_orig;
         return LZMA_BUF_ERROR;
     }
-    if lzma_block_header_encode(block, out.offset(*out_pos as isize)) as c_uint != LZMA_OK as c_uint
+    if lzma_block_header_encode(block, out.offset(*out_pos as isize)) != LZMA_OK
     {
         (*block).filters = filters_orig;
         return LZMA_PROG_ERROR;
@@ -354,7 +354,7 @@ unsafe extern "C" fn block_encode_normal(
     mut out_size: size_t,
 ) -> lzma_ret {
     let ret_: lzma_ret = lzma_block_header_size(block) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     if out_size.wrapping_sub(*out_pos) <= (*block).header_size as size_t {
@@ -379,7 +379,7 @@ unsafe extern "C" fn block_encode_normal(
     };
     let mut ret: lzma_ret =
         lzma_raw_encoder_init(&raw mut raw_encoder, allocator, (*block).filters);
-    if ret as c_uint == LZMA_OK as c_uint {
+    if ret == LZMA_OK {
         let mut in_pos: size_t = 0 as size_t;
         ret = raw_encoder.code.expect("non-null function pointer")(
             raw_encoder.coder,
@@ -394,18 +394,18 @@ unsafe extern "C" fn block_encode_normal(
         );
     }
     lzma_next_end(&raw mut raw_encoder, allocator);
-    if ret as c_uint == LZMA_STREAM_END as c_uint {
+    if ret == LZMA_STREAM_END {
         (*block).compressed_size = (*out_pos)
             .wrapping_sub(out_start.wrapping_add((*block).header_size as size_t))
             as lzma_vli;
         ret = lzma_block_header_encode(block, out.offset(out_start as isize));
-        if ret as c_uint != LZMA_OK as c_uint {
+        if ret != LZMA_OK {
             ret = LZMA_PROG_ERROR;
         }
-    } else if ret as c_uint == LZMA_OK as c_uint {
+    } else if ret == LZMA_OK {
         ret = LZMA_BUF_ERROR;
     }
-    if ret as c_uint != LZMA_OK as c_uint {
+    if ret != LZMA_OK {
         *out_pos = out_start;
     }
     return ret;
@@ -431,7 +431,7 @@ unsafe extern "C" fn block_buffer_encode(
     if (*block).version > 1 as u32 {
         return LZMA_OPTIONS_ERROR;
     }
-    if (*block).check as c_uint > LZMA_CHECK_ID_MAX as c_uint
+    if (*block).check > LZMA_CHECK_ID_MAX as c_uint
         || try_to_compress as c_int != 0 && (*block).filters.is_null()
     {
         return LZMA_PROG_ERROR;
@@ -454,13 +454,13 @@ unsafe extern "C" fn block_buffer_encode(
     if try_to_compress {
         ret = block_encode_normal(block, allocator, in_0, in_size, out, out_pos, out_size);
     }
-    if ret as c_uint != LZMA_OK as c_uint {
-        if ret as c_uint != LZMA_BUF_ERROR as c_uint {
+    if ret != LZMA_OK {
+        if ret != LZMA_BUF_ERROR {
             return ret;
         }
         let ret_: lzma_ret =
             block_encode_uncompressed(block, in_0, in_size, out, out_pos, out_size) as lzma_ret;
-        if ret_ as c_uint != LZMA_OK as c_uint {
+        if ret_ != LZMA_OK {
             return ret_;
         }
     }
