+++ liblzma-rs/src/common/filter_buffer_encoder.rs
@@ -121,7 +121,7 @@ pub unsafe extern "C" fn lzma_raw_buffer_encode(
         set_out_limit: None,
     };
     let ret_: lzma_ret = lzma_raw_encoder_init(&raw mut next, allocator, filters) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     let out_start: size_t = *out_pos;
@@ -138,10 +138,10 @@ pub unsafe extern "C" fn lzma_raw_buffer_encode(
         LZMA_FINISH,
     );
     lzma_next_end(&raw mut next, allocator);
-    if ret as c_uint == LZMA_STREAM_END as c_uint {
+    if ret == LZMA_STREAM_END {
         ret = LZMA_OK;
     } else {
-        if ret as c_uint == LZMA_OK as c_uint {
+        if ret == LZMA_OK {
             ret = LZMA_BUF_ERROR;
         }
         *out_pos = out_start;
